+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
-## [2.1.6-SNAPSHOT] - not released
+## [2.1.6-SNAPSHOT]
 
 ### Added
 
@@ -12,11 +12,24 @@ All notable changes to this project will be documented in this file. This change
 * much more types of `skewness` and `kurtosis`
 * much more effect size functions: `r2-determination`, `eta-sq`, `omega-sq`, `epsilon-sq`, `cohens-f2`, `cohens-q`, `cramers-v`, `cohens-w`, `tschuprows-t`
 * `trim` and `winsor` data based on quantiles
+* distances:
+  - `dist-ang` - angular distance
+  - `sim-cos` - cosine similarity
+  - `angular` - refers to angular distance
 
 ### Fixed
 
 * `moment` didn't work properly for certain cases
 * `effect-size` methods fixes
+* [breaking] fixes around cosine distance and similarity
+
+### Changed
+
+* `fastmath.distance/cosine` - refers to cosine similarity now
+
+### Removed
+
+* [breaking] `fastmath.vector/dist-cos` - wrong implementation
 
 ## [2.1.5]